19102025/osint_tool/requirements.txt
python-whois
dnspython
requests
beautifulsoup4
networkx
matplotlib
Flask
